+++ blink_led_microrl/print.h
@@ -29,38 +29,38 @@ int print_eol();
 // print char
 int print_chr(char c);
 //-----------------------------------------------------------------------------
-// print integer
-int print_int(int i);
+// print long integer
+int print_int(long i);
 //-----------------------------------------------------------------------------
-// print unsigned integer
-int print_uint(unsigned i);
+// print unsigned long integer
+int print_uint(unsigned long i);
 //-----------------------------------------------------------------------------
-// print unsigned integer with '0' on the begin
-int print_uint_ex(unsigned i, char digits);
+// print unsigned long integer with '0' on the begin
+int print_uint_ex(unsigned long i, char digits);
 //-----------------------------------------------------------------------------
-// print binary unsigned integer value
-int print_bin(unsigned i, char digits);
+// print binary unsigned long integer value
+int print_bin(unsigned long i, char digits);
 //-----------------------------------------------------------------------------
-// hex print unsigned integer
-int print_hex(unsigned i, char digits);
+// hex print unsigned long integer
+int print_hex(unsigned long i, char digits);
 //-----------------------------------------------------------------------------
-// print integer as float in NNN.D format [d = (int) (f * 10.)]
-void print_dint(int d);
+// print long integer as float in NNN.D format [d = (int) (f * 10.)]
+void print_dint(long d);
 //-----------------------------------------------------------------------------
-// print "ident = value\r\n" (integer value)
-void print_ival(const char *ident, int i);
+// print "ident = value\r\n" (long integer value)
+void print_ival(const char *ident, long i);
 //-----------------------------------------------------------------------------
-// print "ident = NNN.D\r\n" (integer value as float)
-void print_dval(const char *ident, int d);
+// print "ident = NNN.D\r\n" (long integer value as float)
+void print_dval(const char *ident, long d);
 //-----------------------------------------------------------------------------
-// print "ident = value\r\n" (unsigned integer value)
-void print_uval(const char *ident, unsigned i);
+// print "ident = value\r\n" (unsigned long integer value)
+void print_uval(const char *ident, unsigned long i);
 //-----------------------------------------------------------------------------
-// print "ident = value\r\n" (unsigned bin value)
-void print_bval(const char *ident, unsigned i, char digits);
+// print "ident = value\r\n" (unsigned long bin value)
+void print_bval(const char *ident, unsigned long i, char digits);
 //-----------------------------------------------------------------------------
-// print "ident = value\r\n" (unsigned hex value)
-void print_hval(const char *ident, unsigned i, char digits);
+// print "ident = value\r\n" (unsigned long hex value)
+void print_hval(const char *ident, unsigned long i, char digits);
 //-----------------------------------------------------------------------------
 // print "ident = value\r\n" (string value)
 void print_sval(const char *ident, const char *str);
@@ -75,4 +75,3 @@ void print_flush(void);
 #endif // PRINT_H
 
 /*** end of "print.h" file ***/
-
